app_moviles/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
GDAL==3.4.1
gunicorn==23.0.0
idna==3.8
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.1.0
packaging==24.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
setuptools
djangorestframework-gis
pillow==11.0.0
